Move remaining functions into own classes
