+++ h2/src/main/org/h2/expression/function/SetFunction.java
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2004-2020 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * and the EPL 1.0 (https://h2database.com/html/license.html).
+ * Initial Developer: H2 Group
+ */
+package org.h2.expression.function;
+
+import org.h2.api.ErrorCode;
+import org.h2.engine.SessionLocal;
+import org.h2.expression.Expression;
+import org.h2.expression.ExpressionVisitor;
+import org.h2.expression.Variable;
+import org.h2.message.DbException;
+import org.h2.value.Value;
+
+/**
+ * A SET function.
+ */
+public final class SetFunction extends Function2 {
+
+    public SetFunction(Expression arg1, Expression arg2) {
+        super(arg1, arg2);
+    }
+
+    @Override
+    public Value getValue(SessionLocal session) {
+        Variable var = (Variable) left;
+        Value v = right.getValue(session);
+        session.setVariable(var.getName(), v);
+        return v;
+    }
+
+    @Override
+    public Expression optimize(SessionLocal session) {
+        left = left.optimize(session);
+        right = right.optimize(session);
+        type = right.getType();
+        if (!(left instanceof Variable)) {
+            throw DbException.get(ErrorCode.CAN_ONLY_ASSIGN_TO_VARIABLE_1, left.getTraceSQL());
+        }
+        return this;
+    }
+
+    @Override
+    public String getName() {
+        return "SET";
+    }
+
+    @Override
+    public boolean isEverything(ExpressionVisitor visitor) {
+        if (!super.isEverything(visitor)) {
+            return false;
+        }
+        switch (visitor.getType()) {
+        case ExpressionVisitor.DETERMINISTIC:
+        case ExpressionVisitor.QUERY_COMPARABLE:
+        case ExpressionVisitor.READONLY:
+            return false;
+        case ExpressionVisitor.EVALUATABLE:
+        case ExpressionVisitor.GET_DEPENDENCIES:
+        case ExpressionVisitor.INDEPENDENT:
+        case ExpressionVisitor.NOT_FROM_RESOLVER:
+        case ExpressionVisitor.OPTIMIZABLE_AGGREGATE:
+        case ExpressionVisitor.SET_MAX_DATA_MODIFICATION_ID:
+        case ExpressionVisitor.GET_COLUMNS1:
+        case ExpressionVisitor.GET_COLUMNS2:
+            return true;
+        default:
+            throw DbException.throwInternalError("type=" + visitor.getType());
+        }
+    }
+
+}
